pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "a2a-sdk[sqlite]",
    "pydantic>=2.11.3",
    "pydantic-ai==0.0.46",
    "supabase",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "PyYAML==6.0.2",
    "requests==2.32.3",
    "rich==14.0.0",
    "openai",
    "mcp>=1.4.1",
    "mdurl==0.1.2",
]
